phase 1 of the issue #144/#145 follow-up
